requirements.txt
customtkinter>=5.2.2
qrcode[pil]>=7.4
Pillow>=10.0
python-docx>=1.1.0
lxml>=4.9
